+++ src/pages/task/taskTabs/buildBaronAndAnnotations/Issues/index.ts
@@ -1,0 +1,4 @@
+import Issues from "./Issues";
+import SuspectedIssues from "./SuspectedIssues";
+
+export { SuspectedIssues, Issues };
